Move Op player Event adapter to quest.event package
